+++ app/controllers/dishes_controller.rb
@@ -1,35 +1,6 @@
 class DishesController < ApplicationController
   before_action :set_dish, only: :show
 
-  def show
-    @dish = Dish.find(params[:id])
-    @ingredients = @dish.ingredients
-    @dish_ingredients = @dish.dish_ingredients
-  end
-
-  def increase_amount
-    @dish = Dish.find(params[:dish_id])
-    @ingredients = @dish.ingredients
-    @dish_ingredients = @dish.dish_ingredients
-    @ingredients.each do |ingredient|
-      @dish_ingredients.each do |dish_ingredient|
-        dish_ingredient.base_quantity += ingredient.change_increment
-      end
-    end
-  end
-
-  def decrease_amount
-    @dish = Dish.find(params[:dish_id])
-    @ingredients = @dish.ingredients
-    @dish_ingredients = @dish.dish_ingredients
-    @ingredients.each do |ingredient|
-      @dish_ingredients.each do |dish_ingredient|
-        dish_ingredient.base_quantity -= ingredient.change_increment
-      end
-    end
-  end
-
-
   def create
     @dish = Dish.new(dish_params)
     @restaurant = @dish.restaurant
@@ -40,7 +11,6 @@ class DishesController < ApplicationController
       render :new
     end
   end
-
 
   private
