+++ app/controllers/restaurants_controller.rb
@@ -1,9 +1,5 @@
 class RestaurantsController < ApplicationController
 
-  def index
-    @restaurants = Restaurant.all
-  end
-
   before_action :set_restaurant, only: [:show, :edit, :update, :destroy]
 
   def index
@@ -13,6 +9,9 @@ class RestaurantsController < ApplicationController
   def show
     @restaurant = Restaurant.find(params[:id])
     @dishes = @restaurant.dishes
+    @dishes.each do |dish|
+      @ingredients = dish.ingredients
+    end
   end
 
   def new
@@ -21,7 +20,7 @@ class RestaurantsController < ApplicationController
 
   def create
     @restaurant = Restaurant.new(restaurant_params)
-    @restaurant.user  = current_user
+    @restaurant.user == current_user
     if @restaurant.save
       flash[:success] = "Thanks for adding your restaurant!"
       redirect_to @restaurant
