+++ app/models/dish.rb
@@ -1,6 +1,6 @@
 class Dish < ApplicationRecord
 
-  belongs_to :restaurant, foreign_key: true
+  belongs_to :restaurant
   has_many :dish_ingredients, dependent: :destroy
   has_many :ingredients, through: :dish_ingredients
