+++ app/models/dish_ingredient.rb
@@ -1,6 +1,7 @@
 class DishIngredient < ApplicationRecord
-  belongs_to :dish, foreign_key: true
-  belongs_to :ingredient, foreign_key: true
+  belongs_to :dish
+  belongs_to :ingredient
+  has_many :customized_ingredients, dependent: :destroy
 
   validates_presence_of :base_quantity, :min_quantity, :max_quantity
 end
