+++ app/models/ingredient.rb
@@ -1,6 +1,6 @@
 class Ingredient < ApplicationRecord
 
-  belongs_to :restaurant, foreign_key: true
+  belongs_to :restaurant
   has_many :dish_ingredients, dependent: :destroy
 
   validates_presence_of :name, :unit, :change_increment, :cost, :price
