+++ test/controllers/customized_ingredients_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class CustomizedIngredientsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
